pyproject.toml
[tool.poetry]
name = "trimshare"
version = "0.2.0"
description = ""
authors = ["JonLiuFYI <2997583+JonLiuFYI@users.noreply.github.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.10"
docopt-ng = "^0.7.2"
freezegun = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^22.1.0"
flake8 = "^4.0.1"

[tool.poetry.scripts]
trimshare = 'trimshare:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
